refactor number line exercise, improve mobile
